+++ corehq/apps/userreports/management/commands/find_datasource_mismatches.py
@@ -9,7 +9,7 @@ import pytz
 from django.core.management.base import BaseCommand
 from django.utils.dateparse import parse_datetime
 
-from corehq.apps.userreports.document_stores import get_document_store_for_doc_type
+from corehq.apps.change_feed.data_sources import get_document_store_for_doc_type
 from corehq.apps.userreports.models import get_datasource_config
 from corehq.apps.userreports.util import get_indicator_adapter
 from corehq.util.log import with_progress_bar
