+++ corehq/apps/userreports/management/commands/profile_data_source.py
@@ -6,7 +6,7 @@ import pstats
 
 from django.core.management.base import BaseCommand
 
-from corehq.apps.userreports.document_stores import get_document_store_for_doc_type
+from corehq.apps.change_feed.data_sources import get_document_store_for_doc_type
 from corehq.apps.userreports.models import get_datasource_config
